refactor(site): move app into module
